Add REST endpoint to list `OracleData`

+++ script/generate.sh
@@ -13,6 +13,10 @@ function gen_code() {
 	# generate code, grpc and validators code
 	buf generate
 
+	if [ $? -ne 0 ]; then
+		exit $?
+	fi
+
 	# Since ./proto/github/{grpc-ecosystem,mwitkow} are dependencies,
 	# buf will generate code for them to
 	rm -rf ./github.com
